+++ tutorials/preprocessor/python/README_CN.md
@@ -1,3 +1,4 @@
+[English](README.md) | 中文
 # Preprocessor Python 示例代码
 1. [编译FastDeploy（Python）](../docs/cn/build_and_install), 或直接下载[FastDeploy预编译库（Python）](../docs/cn/build_and_install/download_prebuilt_libraries.md)
 2. 运行示例代码
@@ -6,5 +7,5 @@
 wget https://gitee.com/paddlepaddle/PaddleClas/raw/release/2.4/deploy/images/ImageNet/ILSVRC2012_val_00000010.jpeg
 
 # 运行示例代码
-python preprocess_cvcuda.py
+python preprocess.py --use_cvcuda True
 ```
